+++ packages/eslint-plugin/src/index.ts
@@ -1,4 +1,5 @@
 import "source-map-support/register";
 import rules from "./rules";
+import { defineConfig } from "./rules/check-sql.config";
 
-export = { rules };
+export = { rules, defineConfig };